examples/model-training/pyproject.toml
[tool.poetry]
name = "model-training"
version = "0.1.0"
description = ""
authors = ["Robert Pack <robstar.pack@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
dagster-fusion = { path = "../../python/dagster-fusion", develop = true }
scipy = "^1.8"
scikit-learn = "^1"

[tool.poetry.dev-dependencies]
pytest-datadir = "^1.3.1"
requests-mock = "^1.9.3"
responses = ">=0.21"
